autoconfig and coordinatesDA_within_margin
added autoconfig again: in text use
   autoconfig_AND_perform_mapping(mapping_file_index,mainPath)
instead of
    mapping_file.perform_mapping()

added the function coordinatesDA_within_margin in coordinate_optimization:
there was already a coordinates_within_margin there, but that checks only in one channel and does not remove the other channel as well (leaving you with unmatched donor and acceptor points)

lastly coordinatesDA_within_margin is used in file.find_coordinates.
In perform_mapping it is not necessary, since there the grouping of points is done after margins per channel, and no traces are to be extracted from those points.