Update file.py
the idea is that in file.perform_mapping the initial translation can be either width//2, or a given 2 element array.
But for the BN-TIRF setup the inital estimator needs a large shift, because both channels are not well aligned. Because of that, we need a better initial estimator in file.perform_mapping
the found value overwrites the initial_translation in the config file.